Move definitions of stats to separate files
For #252
