+++ lib/repofetch.rb
@@ -5,6 +5,8 @@ require 'git'
 require 'repofetch/config'
 require 'repofetch/env'
 require 'repofetch/exceptions'
+require 'repofetch/stat'
+require 'repofetch/timespan_stat'
 require 'repofetch/theme'
 require 'repofetch/util'
 
@@ -222,65 +224,6 @@ class Repofetch
       else
         stat_lines.zip(ascii_lines).map(&:reverse)
       end.map { |ascii, stat| [ascii.to_s, stat.to_s] }
-    end
-  end
-
-  # Base class for stats.
-  class Stat
-    attr_reader :label, :value, :emoji
-    attr_writer :theme
-
-    # Creates a stat
-    #
-    # @param [String] label The label of the stat
-    # @param value The value of the stat
-    # @param [String] emoji An optional emoji for the stat
-    def initialize(label, value, emoji: nil)
-      @label = label
-      @value = value
-      @emoji = emoji
-      @label_styles = []
-    end
-
-    def to_s
-      emoji = @emoji
-      emoji = nil unless Repofetch.config.nil? || Repofetch.config.emojis?
-      "#{emoji}#{format_label}: #{format_value}"
-    end
-
-    # Adds a style for the label
-    #
-    # @param [Symbol] style The theme's style to add
-    def style_label!(style)
-      @label_styles << style
-    end
-
-    # Formats the label, including styles.
-    #
-    # @return [String]
-    def format_label
-      return @label if @theme.nil?
-
-      @label_styles.inject(@label) { |label, style| @theme.format(style, label) }
-    end
-
-    # Formats the value
-    #
-    # This simply converts the value to a string, but can be overridden but
-    # subclasses to affect +to_s+.
-    def format_value
-      @value.to_s
-    end
-  end
-
-  # Timespan stat for "x units ago" stats.
-  class TimespanStat < Stat
-    include ActionView::Helpers::DateHelper
-
-    # Formats the value as "x units ago".
-    def format_value(now = nil)
-      now = Time.now if now.nil?
-      "#{distance_of_time_in_words(@value, now)} ago"
     end
   end
